Rework FileMoveIcon into ItemState
Rename FileMoveIcon to ItemState.  But also, eliminate the inner enum in favor of public final static instances.  It makes the syntax much easier, and now we don't have to worry about trying to map backwards from a Status to the object.
